Pokemon Alpha Commit
All moves that have base functionality should work and pokemon can faint to end the battle.
Accuracy Fix.
Effective status message should now appear at the right time.
Attempt to make Pokemon faint, need to reimplement/polish later, works for now.
Todo: make the battlefield a singleton and give it some more cool methods for moves/pokemon to call.
Small fixes.

+++ Game/src/AttackStates/AttackWrapper.java
@@ -47,7 +47,7 @@ public abstract class AttackWrapper extends AttackState {
     }
 
     @Override
-    protected int getAccuracy(Pokemon ourselves, Pokemon opponent) {
+    protected double getAccuracy(Pokemon ourselves, Pokemon opponent) {
         return nextAttack.getAccuracy(ourselves, opponent);
     }
 }
